Fix parseDouble return type

+++ modules/nextflow/src/main/groovy/nextflow/trace/TraceRecord.groovy
@@ -600,7 +600,7 @@ class TraceRecord implements Serializable {
         }
     }
 
-    private long parseDouble( String str, Path file , String row )  {
+    private double parseDouble( String str, Path file , String row )  {
         try {
             str.toDouble()
         }
